добавлена реализация меню

+++ kursach_tp_/friend.h
@@ -5,5 +5,5 @@
 #include <string>
 using namespace std;
 
-string converter(int day, int mouth, int year, double cost, string str);
+string converter(int day, int mouth, int year, int cost, string str);
 void function(string request, string str);
